Move remote gateway packages
