+++ pub/remote_gateway.go
@@ -6,7 +6,6 @@ import (
 	"fmt"
 	"net/http"
 	"net/url"
-	"os/exec"
 
 	"github.com/gin-gonic/gin"
 	"github.com/gorilla/websocket"
@@ -306,162 +305,4 @@ func (l *remoteLifecycleClient) AuthorizeCaller(jobName, jobVersion, endpoint st
 	}
 
 	return response.JobCallAuthData, nil
-}
-
-func remoteGatewayEndpoint(c *gin.Context, cfg *Config, jobPath string) {
-	requestId := getRequestTracingId(c.Request, cfg.RequestTracingHeader)
-	logger := log.New(log.Ctx{
-		"requestId": requestId,
-	})
-
-	logger.Info("Incoming forwarding request from main Pub", log.Ctx{"method": c.Request.Method, "path": c.Request.URL.Path})
-	statusCode, err := handleRemoteGatewayRequest(c, cfg, logger, requestId, jobPath)
-	if err != nil {
-		errorStr := err.Error()
-		logger.Error("Proxy request error", log.Ctx{
-			"status": statusCode,
-			"error":  errorStr,
-			"path":   c.Request.URL.Path,
-		})
-		c.JSON(statusCode, gin.H{
-			"error":     fmt.Sprintf("Proxy request error: %s", errorStr),
-			"status":    http.StatusText(statusCode),
-			"requestId": requestId,
-		})
-	}
-}
-
-func handleRemoteGatewayRequest(
-	c *gin.Context,
-	cfg *Config,
-	logger log.Logger,
-	requestId string,
-	jobPath string,
-) (int, error) {
-	if !cfg.RemoteGatewayMode {
-		return http.StatusUnauthorized, errors.New("Forwarding endpoint is only available in remote gateway mode")
-	}
-
-	if c.Request.Method != "POST" && c.Request.Method != "GET" {
-		c.Writer.Header().Set("Allow", "GET, POST")
-		return http.StatusMethodNotAllowed, errors.New("Method not allowed")
-	}
-
-	jobName := c.Param("job")
-	if jobName == "" {
-		return http.StatusBadRequest, errors.New("Couldn't extract job name")
-	}
-	jobVersion := c.Param("version")
-	if jobVersion == "" {
-		return http.StatusBadRequest, errors.New("Couldn't extract job version")
-	}
-
-	gatewayToken := c.Request.Header.Get(RemoteGatewayTokenHeader)
-	if gatewayToken == "" {
-		return http.StatusUnauthorized, errors.Errorf("Pub gateway expects token in %s header", RemoteGatewayTokenHeader)
-	}
-	if gatewayToken != cfg.RemoteGatewayToken {
-		return http.StatusUnauthorized, errors.New("Pub gateway token is invalid")
-	}
-
-	jobInternalName := c.Request.Header.Get(JobInternalNameHeader)
-	if jobInternalName == "" {
-		return http.StatusBadRequest, errors.Errorf("Pub gateway expects job name in %s header", JobInternalNameHeader)
-	}
-
-	job := &JobDetails{
-		Name:         jobName,
-		Version:      jobVersion,
-		InternalName: jobInternalName,
-	}
-	callerName := c.Request.Header.Get(cfg.CallerNameHeader)
-
-	urlPath := JoinURL("/pub/job/", job.Name, job.Version, jobPath)
-	targetUrl := TargetURL(cfg, job, urlPath)
-
-	logger.Debug("Forwarding request to job", log.Ctx{
-		"jobName":         jobName,
-		"jobVersion":      jobVersion,
-		"jobInternalName": jobInternalName,
-		"targetUrl":       targetUrl.String(),
-	})
-
-	ServeReverseProxy(targetUrl, c, job, cfg, logger, requestId, callerName)
-	return http.StatusOK, nil
-}
-
-// A command called remotely by Racetrack host to manage infrastructure target
-func remoteCommandEndpoint(c *gin.Context, cfg *Config) {
-	requestId := getRequestTracingId(c.Request, cfg.RequestTracingHeader)
-	logger := log.New(log.Ctx{
-		"requestId": requestId,
-	})
-
-	logger.Info("Incoming remote command", log.Ctx{"method": c.Request.Method, "path": c.Request.URL.Path})
-	statusCode, err := handleRemoteCommandRequest(c, cfg, logger, requestId)
-	if err != nil {
-		errorStr := err.Error()
-		logger.Error("Remote command error", log.Ctx{
-			"status": statusCode,
-			"error":  errorStr,
-			"path":   c.Request.URL.Path,
-		})
-		c.JSON(statusCode, gin.H{
-			"error":     fmt.Sprintf("Remote command: %s", errorStr),
-			"status":    http.StatusText(statusCode),
-			"requestId": requestId,
-		})
-	}
-}
-
-type remoteCommandRequest struct {
-	Command string `json:"command"`
-	Workdir string `json:"workdir"`
-}
-
-func handleRemoteCommandRequest(
-	c *gin.Context,
-	cfg *Config,
-	logger log.Logger,
-	requestId string,
-) (int, error) {
-	gatewayToken := c.Request.Header.Get(RemoteGatewayTokenHeader)
-	if gatewayToken == "" {
-		return http.StatusUnauthorized, errors.Errorf("remote Pub expects token in %s header", RemoteGatewayTokenHeader)
-	}
-	if gatewayToken != cfg.RemoteGatewayToken {
-		return http.StatusUnauthorized, errors.New("remote Pub token is invalid")
-	}
-
-	var request remoteCommandRequest
-	err := c.BindJSON(&request)
-	if err != nil {
-		return http.StatusBadRequest, errors.Wrap(err, "failed to parse request data as JSON")
-	}
-	if request.Command == "" {
-		return http.StatusBadRequest, errors.New("command field is empty")
-	}
-
-	logger.Debug("Executing remote command", log.Ctx{
-		"command": request.Command,
-	})
-	cmd := exec.Command("sh", "-c", request.Command)
-	if request.Workdir != "" {
-		cmd.Dir = request.Workdir
-	}
-	output, err := cmd.CombinedOutput()
-	if err != nil {
-		logger.Error("Command failed", log.Ctx{
-			"command":   request.Command,
-			"exit_code": cmd.ProcessState.ExitCode(),
-			"err":       err,
-		})
-	}
-
-	c.JSON(http.StatusOK, gin.H{
-		"output":    string(output),
-		"exit_code": cmd.ProcessState.ExitCode(),
-		"requestId": requestId,
-	})
-	return http.StatusOK, nil
 }
